feat: created composable for uiState
