--- packages/core/nuxt-theme-module/theme/assets/ui-state.ts
@@ -1,54 +1,0 @@
-import Vue from 'vue';
-import VueCompositionAPI, { reactive, computed } from '@vue/composition-api';
-
-// We need to register it again because of Vue instance instantiation issues
-Vue.use(VueCompositionAPI);
-
-const state = reactive({
-  isCartSidebarOpen: false,
-  isWishlistSidebarOpen: false,
-  isLoginModalOpen: false,
-  isCategoryGridView: true,
-  isFilterSidebarOpen: false
-});
-
-// TODO: https://github.com/DivanteLtd/vue-storefront/issues/4876
-const isCartSidebarOpen = computed(() => state.isCartSidebarOpen);
-const toggleCartSidebar = () => {
-  state.isCartSidebarOpen = !state.isCartSidebarOpen;
-};
-
-const isWishlistSidebarOpen = computed(() => state.isWishlistSidebarOpen);
-const toggleWishlistSidebar = () => {
-  state.isWishlistSidebarOpen = !state.isWishlistSidebarOpen;
-};
-
-const isLoginModalOpen = computed(() => state.isLoginModalOpen);
-const toggleLoginModal = () => {
-  state.isLoginModalOpen = !state.isLoginModalOpen;
-};
-
-const isCategoryGridView = computed(() => state.isCategoryGridView);
-const toggleCategoryGridView = () => {
-  state.isCategoryGridView = !state.isCategoryGridView;
-};
-
-const isFilterSidebarOpen = computed(() => state.isFilterSidebarOpen);
-const toggleFilterSidebar = () => {
-  state.isFilterSidebarOpen = !state.isFilterSidebarOpen;
-};
-
-const uiState = {
-  isCartSidebarOpen,
-  isWishlistSidebarOpen,
-  isLoginModalOpen,
-  isCategoryGridView,
-  isFilterSidebarOpen,
-  toggleCartSidebar,
-  toggleWishlistSidebar,
-  toggleLoginModal,
-  toggleCategoryGridView,
-  toggleFilterSidebar
-};
-
-export default uiState;
