+++ packages/core/nuxt-theme-module/theme/components/LoginModal.vue
@@ -138,9 +138,7 @@ import { SfModal, SfInput, SfButton, SfCheckbox, SfLoader, SfAlert } from '@stor
 import { ValidationProvider, ValidationObserver, extend } from 'vee-validate';
 import { required, email } from 'vee-validate/dist/rules';
 import { useUser } from '<%= options.generate.replace.composables %>';
-import uiState from '~/assets/ui-state';
-
-const { isLoginModalOpen, toggleLoginModal } = uiState;
+import { useUiState } from '~/composables';
 
 extend('email', {
   ...email,
@@ -165,6 +163,7 @@ export default {
     ValidationObserver
   },
   setup() {
+    const { isLoginModalOpen, toggleLoginModal } = useUiState();
     const form = ref({});
     const isLogin = ref(false);
     const createAccount = ref(false);
