+++ packages/core/nuxt-theme-module/theme/composables/index.ts
@@ -1,5 +1,7 @@
 import useUiHelpers from './useUiHelpers';
+import useUiState from './useUiState';
 
 export {
-  useUiHelpers
+  useUiHelpers,
+  useUiState
 };
